% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.mat.R
\name{combine.mat}
\alias{combine.mat}
\title{Combine matrices with different dimensions, add NA for additional rows/columns
Used in outputting the DTR tree.}
\usage{
combine.mat(m1, m2, by = "column")
}
\arguments{
\item{m1}{Matrix 1.}

\item{m2}{Matrix 2.}
}
\description{
Combine matrices with different dimensions, add NA for additional rows/columns
Used in outputting the DTR tree.
}
