% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.lambda.R
\name{CV.lambda}
\alias{CV.lambda}
\title{Cross-validation to decide minimum purity improvement for splitting.}
\usage{
CV.lambda(
  Y,
  A,
  H,
  g.opt,
  lambda.pct,
  K = 10,
  m.method = c("AIPW", "randomForest"),
  depth = 5,
  minsplit = 20
)
}
\arguments{
\item{Y}{A vector of outcome of interest.}

\item{A}{Treatment vector.}

\item{H}{A matrix of covariates before assigning final treatment, excluding previous treatment variables.}

\item{g.opt}{Optimal dynamic treatment regime.}

\item{K}{Number of folds.}

\item{m.method}{Method for calculating estimated conditional mean.}

\item{depth}{Maximum tree depth.}

\item{minsplit}{Minimal node size.}
}
\description{
Cross-validation to decide minimum purity improvement for splitting.
}
