% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M.propen.R
\name{M.propen}
\alias{M.propen}
\title{Estimate propensity score by a multinomial model.}
\usage{
M.propen(A, Xs)
}
\arguments{
\item{A}{Treatment vector.}

\item{Xs}{Covariate matrix.}
}
\description{
Estimate propensity score by a multinomial model.
}
