% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mus.AIPW.R
\name{mus.AIPW}
\alias{mus.AIPW}
\title{Calculate weighted mean outcome with AIPW.}
\usage{
mus.AIPW(Y, A, pis.hat, mus.reg)
}
\arguments{
\item{Y}{A vector of outcome of interest.}

\item{A}{Treatment vector.}

\item{pis.hat}{Estimated propensity score matrix.}

\item{mus.reg}{Regression-based conditional mean outcome.}
}
\description{
Calculate weighted mean outcome with AIPW.
}
