% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Opt.A.R
\name{Opt.A}
\alias{Opt.A}
\title{Choose the optimal treatment with given pseudo outcome matrix in one given stage.}
\usage{
Opt.A(A, mus.hat)
}
\arguments{
\item{A}{Treatment options given in this stage}

\item{mu.hat}{The final counterfactural outcome according to the treatment.}
}
\description{
Choose the optimal treatment with given pseudo outcome matrix in one given stage.
}
