% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.DTR.R
\name{predict.DTR}
\alias{predict.DTR}
\title{Predit optimal treatment using the output from DTRtree.}
\usage{
\method{predict}{DTR}(treeout, newdata)
}
\arguments{
\item{treeout}{An object outputted from DTRtree function.}

\item{newdata}{New data containing H (history).}
}
\description{
Predit optimal treatment using the output from DTRtree.
}
