% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_DTR.R
\name{predict_DTR}
\alias{predict_DTR}
\title{Predit optimal treatment using the output from DTRtree.}
\usage{
predict_DTR(treeout, newdata)
}
\arguments{
\item{treeout}{An object outputted from DTRtree function.}

\item{newdata}{New data containing H (history).}
}
\description{
Predit optimal treatment using the output from DTRtree.
}
