% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reg.mu.R
\name{Reg.mu}
\alias{Reg.mu}
\title{Estimate conditional means for multiple stages}
\usage{
Reg.mu(Y, As, H)
}
\arguments{
\item{Y}{A continous outcome of interest.}

\item{As}{(A1, A2, ...) a matrix of treatments at multiple stages; Stage t has treatment K_t options labeled as 0, 1, ..., K_t-1.}

\item{H}{A matrix of covariates before assigning final treatment, excluding previous treatment variables.}
}
\description{
Estimate conditional means for multiple stages
}
