% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Split.X.lh.R
\name{Split.X.lh}
\alias{Split.X.lh}
\title{Search for best split by lookahead
Compare single best split with further split to calculate the final cutoff.}
\usage{
Split.X.lh(X, A, H, mus.hat, minsplit = 20)
}
\arguments{
\item{X}{Covariate matrix.}

\item{A}{A vector of observed treatment options.}

\item{H}{A matrix of covariates before assigning final treatment, excluding previous treatment variables.}

\item{mus.hat}{Estimated conditional mean outcome.}

\item{minsplit}{Minimal node size.}
}
\description{
Search for best split by lookahead
Compare single best split with further split to calculate the final cutoff.
}
