% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Split.X.R
\name{Split.X}
\alias{Split.X}
\title{Split a parent node into two child nodes by a covariate.
Split the parent node in order to maximize the outcome of interest by signing different treatment in child nodes. Calculate new outcome means for each child node.}
\usage{
Split.X(X, A, mus.hat, minsplit = 20)
}
\arguments{
\item{X}{Covariate matrix.}

\item{A}{A vector of observed treatments.}

\item{mus.hat}{Estimated conditional mean outcome.}

\item{minsplit}{Minimal node size.}
}
\description{
Split a parent node into two child nodes by a covariate.
Split the parent node in order to maximize the outcome of interest by signing different treatment in child nodes. Calculate new outcome means for each child node.
}
